fix the pbs-install command to work on macOS
MacOS ships with an ancient Bash which does not support various
features including not supporting the `mapfile` builtin.

Add function `read_array_to_file` to replicate `mapfile`'s ability
to read a file into an array but just using the `read` builtin.

Add function `sort_version_list` to sort a list of versions into
a second array instead of trying to do it with a mix of IFS and a
call to `sort`.